Medf 17 flujo autenticacion

+++ entity/userRol.go
@@ -7,10 +7,10 @@ func (UserRol) TableName() string {
 }
 
 type UserRol struct {
-	UsuarioRolID uint      `gorm:"primaryKey;autoIncrement;column:re-usuario-rol_id"`
-	UsuarioID    *uint     `gorm:"column:usuario_id" json:"usuario_id"`
-	RolID        *uint     `gorm:"column:rol_id" json:"rol_id"`
-	EntidadID    *uint     `gorm:"column:entidad_id" json:"entidad_id"`
+	UsuarioRolID int64     `gorm:"primaryKey;autoIncrement;column:re-usuario-rol_id"`
+	UsuarioID    *int64    `gorm:"column:usuario_id" json:"usuario_id"`
+	RolID        *int64    `gorm:"column:rol_id" json:"rol_id"`
+	EntidadID    *int64    `gorm:"column:entidad_id" json:"entidad_id"`
 	FechaInicio  time.Time `gorm:"column:fecha_inicio" json:"fecha_inicio"`
 	FechaFin     time.Time `gorm:"column:fecha_fin" json:"fecha_fin"`
 	Indefinido   bool      `gorm:"column:indefinido" json:"indefinido"`
